+++ cookbooks/vm/recipes/vagrant.rb
@@ -1,15 +1,33 @@
 
-node.set['vagrant']['url'] = 'https://releases.hashicorp.com/vagrant/1.9.3/vagrant_1.9.3_x86_64.deb'
-node.set['vagrant']['checksum'] = 'faff6befacc7eed3978b4b71f0dbb9c135c01d8a4d13236bda2f9ed53482d2c4'
+vagrant_version = '1.9.3'
+vagrant_deb_file = "vagrant_#{vagrant_version}_x86_64.deb"
+vagrant_checksum = 'faff6befacc7eed3978b4b71f0dbb9c135c01d8a4d13236bda2f9ed53482d2c4'
 
-include_recipe 'vagrant'
+vagrant_plugins = {
+  'vagrant-cachier' => '1.2.1',
+  'vagrant-berkshelf' => '5.1.1',
+  'vagrant-omnibus' => '1.5.0',
+  'vagrant-toplevel-cookbooks' => '0.2.4',
+  'vagrant-lxc' => '1.2.3'
+}
 
-install_vagrant_plugin 'vagrant-cachier', '1.2.1'
-install_vagrant_plugin 'vagrant-berkshelf', '5.1.1'
-install_vagrant_plugin 'vagrant-omnibus', '1.5.0'
-install_vagrant_plugin 'vagrant-toplevel-cookbooks', '0.2.4'
-install_vagrant_plugin 'vagrant-lxc', '1.2.3'
+# download and install vagrant
+remote_file "#{Chef::Config[:file_cache_path]}/#{vagrant_deb_file}" do
+  source "https://releases.hashicorp.com/vagrant/#{vagrant_version}/#{vagrant_deb_file}"
+  checksum vagrant_checksum
+  notifies :install, 'dpkg_package[vagrant]', :immediately
+end
+dpkg_package 'vagrant' do
+  source "#{Chef::Config[:file_cache_path]}/#{vagrant_deb_file}"
+  version vagrant_version
+end
 
+# install vagrant plugins
+vagrant_plugins.each do |name, version|
+  install_vagrant_plugin(name, version)
+end
+
+# set default provider
 bashd_entry 'set-vagrant-default-provider' do
   user devbox_user
   content "export VAGRANT_DEFAULT_PROVIDER=#{vmware? ? 'virtualbox' : 'docker'}"
@@ -21,7 +39,6 @@ end
 %w(lxc lxc-templates cgroup-lite redir bridge-utils).each do |pkg|
   package pkg
 end
-
 bash 'add vagrant-lxc sudoers permissions' do
   environment devbox_user_env
   code 'vagrant lxc sudoers'